refactor(policy): extract git-remote parsing into policy/_remote.py
Pre-existing verification sweep caught the CI file-length guard: this
PR's additions pushed discovery.py from 2068 to 2252 lines, over the
2100-line ratchet. Extract the cohesive git-remote identity family --
_git_remote_origin_url, _remote_url_parts, _parse_remote_url,
_extract_org_host_port_from_git_remote, _extract_org_from_git_remote --
into a new policy/_remote.py (discovery.py back to ~2091, re-exports the
names so callers/tests are unchanged).

This also makes the remote-origin owner-guard cleaner: _remote.py is a
dedicated, single-decision module, so the guard binds to its own
owner-registry record (selector free) instead of piggy-backing on
discovery.py's cached-policy-shape record. Guard, mutation case, and the
tests that patch the moved internals (subprocess/urlparse/_parse_remote_url)
are re-pointed at _remote.
